+++ cuvis_ai_adaclip/examples_cuvis/experiments/analyze_thresholds.py
@@ -28,11 +28,11 @@ try:
 except ImportError:
     HAS_SEABORN = False
 
-from cuvis_ai.node.band_selection import CIRFalseColorSelector
+from cuvis_ai.node.channel_selector import CIRSelector
 from cuvis_ai.node.data import LentilsAnomalyDataNode
 from cuvis_ai_core.data.datasets import SingleCu3sDataModule
 from cuvis_ai_core.pipeline.pipeline import CuvisPipeline
-from cuvis_ai_core.utils.types import ExecutionStage
+from cuvis_ai_schemas.enums import ExecutionStage
 
 from cuvis_ai_adaclip import (
     AdaCLIPDetector,
@@ -104,7 +104,7 @@ def main(**kwargs) -> None:
     data_node = LentilsAnomalyDataNode(
         normal_class_ids=[0, 1],
     )
-    band_selector = CIRFalseColorSelector(nir_nm=nir_nm, red_nm=red_nm, green_nm=green_nm)
+    band_selector = CIRSelector(nir_nm=nir_nm, red_nm=red_nm, green_nm=green_nm)
 
     use_half_precision = kwargs.get("use_half_precision", True)
     enable_warmup = kwargs.get("enable_warmup", True)